+++ pymongo/periodic_executor.py
