+++ tools/synchro.py
@@ -108,7 +108,7 @@ replacements = {
     "async_set_fail_point": "set_fail_point",
     "async_ensure_all_connected": "ensure_all_connected",
     "async_repl_set_step_down": "repl_set_step_down",
-    "AsyncPeriodicExecutor(": "PeriodicExecutor(",
+    "AsyncPeriodicExecutor": "PeriodicExecutor",
     "async_wait_for_event": "wait_for_event",
 }
